[WPT] Fix currentScale-change-repaint test
The test was scaling a 100x100 by 0.5 and trying to match 100x100
reference. Fix the initial size.

And it was comparing to the svg ref instead of the html ref, resulting in different padding.

And also make it use reftest-wait to be sure.

Fixed: 1368167
Change-Id: I6349b83bb8a6d64e8d4499e22c0f9c01ba5b14ef
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3919130
Commit-Queue: Stephen Chenney <schenney@chromium.org>
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051529}