Restructure tests to relevant base and concrete class.
In order to follow good testing principles, and the DRY coding
principle, strictly test the public methods in their respective classes.

In this CL, we move the relevant tests from the derived class to the
base class which primarily test the base classes public methods.
We are left with a much cleaner testing structure moving forward.

BUG=chromium:1325220

Change-Id: I5d0365409173681e488ac9e7dc71556621dc1f96
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3647149
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Commit-Queue: Hirthanan Subenderan <hirthanan@google.com>
Cr-Commit-Position: refs/heads/main@{#1051535}