Switch Privacy Sandbox Ads APIs flag away from --enable-features
Currently, it invokes --enable-features which means that any other usage
of that switch is ignored. Instead, we should move to using
switch-dependent feature overrides so that both can be used
simultaneously.

As one of the features is outside the blink and content layers, we add
an equivalent chrome-layer feature override mechanism.

Bug: 1366912
Change-Id: Ia86e86e719ed9606a6488122196bc5ca7dab5421
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3913666
Reviewed-by: John Delaney <johnidel@chromium.org>
Commit-Queue: Alex Turner <alexmt@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051554}