inputs: Update PK candidates asynchronously for System PK.
We found that a significant part of the typing latency comes from
updating the candidates window. When Ash is updating and
rerendering the candidates, it blocks the handling of the current
key event.

Move candidates window rendering out of the critical path by
running it asynchronously. Use a OneShotTimer with 0 second delay
as it is now the preferred pattern for this. This will not create
'lag' for the candidates window as the candidates will simply update
in the next run loop, which is not noticeable by users.

Before this CL, the mean key event latency as measured by Tast were:
5.58, 5.49, 5.70 over three runs. After this CL, they reduced to:
4.59, 4.40, 4.57 over three runs. This is an improvement of almost
1ms (15-20%).

Bug: b/241994527
Change-Id: I744a1e91ed95026ffe118c31109a5563fba33e42
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3916617
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Jing Wang <jiwan@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051618}