[ChromeCart] Optimize domain skipping logic
In https://crrev.com/c/3211685, we have routed a boolean from browser
process to renderer to indicate whether renderer should skip operations
for this domain. The request to get this boolean is currently initiated
from two methods in CommerceHintAgent, namely DidStartNavigation and
DidFinishLoad. The boolean is then cached in CommerceHintAgent and
used by other methods when they need to decide whether to skip
operations on current domain.

The problem about this approach is that some functions are called
before the boolean gets passed to browser process, so they are
always triggered until the boolean is initialized. To fix this, this
CL makes the cached boolean a optional value, so that we also skip
operations if the optional value has not been initialized. We expect
this change to be neutral to the accuracy of commerce hint detection,
because we don't expect operations to happen that early in a
renderer's lifecycle (i.e. meaningful events should happen after
the boolean gets initialized). One exception is
DidCommitProvisionalLoad call which could happen very early, and
therefore we add logic to initiate another skip boolean request
there.

Bug: 1366589
Change-Id: Id6661b9fc3d9ce9e2aefa16b086a89460d791b77
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3911548
Commit-Queue: Yue Zhang <yuezhanggg@chromium.org>
Reviewed-by: Wei-Yin Chen <wychen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051656}