[CrOS Cellular] Add connected_apn property to ManagedCellularProperties
The last good APN is a property set by Shill, that only indicates the
last APN used by Shill to connect to a network. To avoid additional
logic in the revamped APN Settings UI, we will provide a property that
is only populated if the Cellular network is connected, and a last good
APN is reported by Shill.

Similar to the last good APN property, it is only read by Chrome, and it
is not required to be sent back to Shill.

Bug: b:162365553

Change-Id: Icfb5a5caa2fbb04a2b31d652052c63973220e308
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3890976
Reviewed-by: Tom Sepez <tsepez@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Gordon Seto <gordonseto@google.com>
Commit-Queue: Emmanuel Arias Soto <eariassoto@google.com>
Cr-Commit-Position: refs/heads/main@{#1051660}