Refactor PreferredAppUpdateWaiter test utility
This CL updates PreferredAppUpdateWaiter to work for synchronous and
asynchronous updates to the preferred app database.
This also adds a helper function SetSupportedLinksPreferenceAndWait()
that combines setting an app as preferred and waiting for the
preference to propagate through to the current process' preference
database.

Bug: b:248979304
Change-Id: If8855f8c9ba967cdd2cc3cafe81cb0e59ec3e112
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3920454
Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
Reviewed-by: Maggie Cai <mxcai@chromium.org>
Commit-Queue: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051695}