[A11Y] Give focus to "block" after "allow" in Android permission dialog
When using TalkBack, if a permission prompt is triggered, the "allow"
button is focused by the screen reader as expected. If the user swipes
right to go to the next element, the focus should move to "block", not
the "home" button as it does now. This CL fixes the order.

Fixed: 1174212
Change-Id: I1b14f77746c22fa5bafbb3aebc853260041b4faa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3914415
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Andy Paicu <andypaicu@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051707}