[TangibleSync] Only show Tangible Sync view for specific access points
Currently tangible sync view is shown for access points that are not
supposed to show it (e.g., signin promos). This is because
SyncConsentFragmentBase checks for the flag and decides whether to
show tangible sync view or not.

Since tangible sync flow is a two step process TangibleSyncCoordinator
should decide when to show tangible sync view. This cl adds two
methods to SyncConsentActivityLauncher and SyncConsentFragmentBase
to specifically create fragment arguments for tangible sync view.
This way only TangibleSyncCoordinator can decide when to show
tangible sync view.

Bug: 1366302
Change-Id: I6c352454bd1134a0ee57514e21cbc932de270d69
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3909652
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Commit-Queue: Tanmoy Mollik <triploblastic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051775}