Dictation: Add refresh() API to ParseStrategy
This change adds a refresh() API to ParseStrategy to avoid destroying
ParseStrategys when the SpeechParser is refreshed. This will be
important because the PumpkinParseStrategy will manage a sandboxed
iframe (implementation in a downstream change) and we'll want to keep
it alive, instead of destroying and recreating it multiple times.

AX-Relnotes: N/A
Bug: 1258190
Change-Id: I3f2b83da65b2348b7f6a71e2280b8ab73996947e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3913730
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Reviewed-by: Katie Dektar <katie@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051845}