Shared Storage: Add JS API to get remaining budget for selectURL.
We add to the shared storage worklet JS API the method
`sharedStorage.remainingBudget()`, which returns a double representing
the remaining bits in the origin's budget for `selectURL()`. We also
add a timing histogram to keep track of how long the calls to this
method take.

Bug: 1218540,1366067
Change-Id: Idadb11084570fcd8d4a4e856918450a8f7364676
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3852345
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Cammie Smith Barnes <cammie@chromium.org>
Reviewed-by: Will Harris <wfh@chromium.org>
Reviewed-by: Yao Xiao <yaoxia@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051895}