[Android][UPM] Evict users based on the finch configured error lists
This CL replaces hardcoded specification of the unrecoverable GMS Core
backend error codes with the finch configurable lists of ignored and
retriable errors.

BAD_REQUEST and DEVELOPER_ERROR codes are added to the default list
of ignored errors to keep the existing behaviour.

Authentication errors are not ignored by default and will lead to
eviction instead of asking the user to sign in.
Adding AUTH_ERROR_RESOLVABLE or AUTH_ERROR_UNRESOLVABLE to the ignored
errors list will result in launching the resolution UI for the
corresponding errors.

Both resolvable and unresolvable errors have kAuthErrorResolvable
error type which triggers the same UI. However, after
crrev.com/c/3914686 these errors will cause the proxy backend to not
return the result and fallback the operation if the recovery type is
not kRecoverable.

Bug: 1362464
Change-Id: Iea70fba3fde3d9f3bfc0708a22eee88cda0bade2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3913390
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Vasilii Sukhanov <vasilii@chromium.org>
Commit-Queue: Maxim Anufriev <maxan@google.com>
Cr-Commit-Position: refs/heads/main@{#1051921}