CloseAll: Allow toasts to persist on hover
This CL adds the ability to make toasts persist when the mouse is
hovering over them. It does this by moving responsibility for handling
toast expiration to the ToastOverlay and using a OneShotTimer to allow
for us to stop and reset the timer when the mouse is hovering over the
toast. We detect hover events through the ToastHoverObserver class.

We also add a unit test to toast_manager_unittest.cc to verify this new
behavior.

Bug: b:247953523
Change-Id: If9a40ab7016838be8c5d508b3998e8e33d0194c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3906498
Commit-Queue: Ben Becker <benbecker@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Reviewed-by: Andre Le <leandre@chromium.org>
Reviewed-by: Alex Newcomer <newcomer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051932}