Experiment with not cancelling search prefetch requests
This CL prevents the cancellation logic in search prefetch from
cancelling ongoing prefetches and prerenders when the omnibox no longer
has the specific suggestion in it. This simplifies the serving logic by
moving from 3 entry points around navigation (marked as clicked, cancel,
serve) to 1 entry point (serve). If this feature is launchable, we will
be able to clean up a lot of this code as well.

Bug: 1361590
Change-Id: I848e6a1fe5942f440157b360af10a5db9357ce61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3915270
Reviewed-by: Simon Pelchat <spelchat@chromium.org>
Commit-Queue: Ryan Sturm <ryansturm@chromium.org>
Reviewed-by: Lingqi Chi <lingqi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1051987}