Dictation: Move Pumpkin install request to PumpkinParseStrategy
This change moves pumpkin installation logic to PumpkinParseStrategy.
This will help simplify plumbing when we spin up a
SandboxedPumpkinTagger from PumpkinData.

This change also unflakes the DictationPumpkinInstallTest.WaitForInstall
test. The reason why the test was flaking is because the ScopedTempDirs
were being destroyed before the files could be read. The fix is to
ensure that they stay alive until the test fixture tears down.

AX-Relnotes: N/A
Bug: 1258190, 1367156
Change-Id: I404bf651998ae1704537a5a356559aaa740f4f04
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3914314
Reviewed-by: Katie Dektar <katie@chromium.org>
Commit-Queue: Akihiro Ota <akihiroota@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052025}