Clean up OverlayCandidate uv rect clipping
OverlayCandidate::ApplyClip() is only used by OverlayCandidateFactory,
so move it to an anonymous namespace in that .cc file.

Move the |uv_rect| transformation to ApplyClip(), so that it can be
skipped if there's no clip to apply.

ApplyClip() can be called potentially three times, so batch these
clips together by intersection, and apply once.

Bug: 1357882
Change-Id: Ibb4436cdee10283174ebc30f66a5fdc4a6070a8a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3913456
Reviewed-by: Peter McNeeley <petermcneeley@chromium.org>
Commit-Queue: River Gilhuly <rivr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052040}