[WebContentsImpl] Don't call didFinishNavigation in child frames
Part of reducing navigation cost to reduce scroll jank. Java observers
don't do work when didFinishNavigation is called in child frames,
so we don't have to notify java in this case.

The changes to restrict didFinishNavigation to main frames is not behind a flag, so to measure the effect on jank, we added a didFinishNavigationNoop function (behind a flag) to mirror what didFinishNavigation did before this change. We think that the overhead from overrides add up to something measurable, so I added a redundant boolean check to make sure this override doesn't get optimized (might get optimized if it's empty but not completely sure). I also wanted it to be as close as possible to what it was doing before, so that we can capture more accurately the effect on jank metrics.

Bug: 1359110
Change-Id: Ib431a9f0b2bfe238ab1077e7fcebca064e3ce27e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3915284
Reviewed-by: Richard Coles <torne@chromium.org>
Commit-Queue: Peilin Wang <peilinwang@google.com>
Reviewed-by: Ella Ge <eirage@chromium.org>
Reviewed-by: Colin Blundell <blundell@chromium.org>
Reviewed-by: Yaron Friedman <yfriedman@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052051}