[remoting] Move application windows during X11 desktop-resize.
When resizing a monitor, other monitor's offsets are adjusted to
re-pack the layout. This CL tries to move any application windows so
their relative position is kept with their monitor.

The algorithm here is simplistic and only looks at each window's
top-left corner. An improvement might be to measure the area of
overlap between each window and each Output, and have the window
track whichever Output it overlaps most with?

Bug: 1366943
Change-Id: I824780f647b2a3c3580afab0b00dc68ef5be2daf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3914389
Commit-Queue: Lambros Lambrou <lambroslambrou@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052062}