Read/write cache with the FPS cache clearing filter info
On writing/updating cache, store an ID to HttpResponseInfo.
On reading cache, check whether the site of the request matches the
cache clearing filter. If so, doom the entry and restart the
transaction if:
- it does not have an ID written previously, or
- it has the ID but is smaller than the ID in the filter map (which
means the cache was written in previous browser runs and should be
bypassed).

Bug: 1316090
Change-Id: I94b367107d0d6ab32197f71468b118f06d015d9d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3907968
Commit-Queue: Shuran Huang <shuuran@chromium.org>
Reviewed-by: Maks Orlovich <morlovich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052083}