Create Reselect Button for Chrome Media Picker
Creates a "Reselect" button that the Chrome Media Picker can display
when a given media list uses a delegated source list. This button allows
the user to change their selection in a delegated source list by
creating a discoverable way for the user to re-summon the system picker.

At present, only one of the Reselect or Audio buttons can be shown;
however, future UI changes to the audio button will move it to its own
row where both buttons could be shown. At present, the media lists that
use a delegated source list do not attempt/have the ability to share
audio. In the medium term, it is likely that if that capability were
added, such a choice would also end up delegated.

An example of how this looks on the "Screen" pane (though it is also
visible on the "Window" pane) can be found at:
https://storage.cloud.google.com/chromium-translation-screenshots/248512b84685691fb1fece7c4f29251f021cbce7

Bug: 1351570
Change-Id: I8c015263e7c7239abacda0a88cb9778fde469fc2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3906373
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Elad Alon <eladalon@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052087}