infra: Bump expose_merge_script_failures to 10%
This bumps the expose_merge_script_failures experiment on all trybots
from 1% to 10%.

And also bumps it on chromeos-amd64-generic-rel from 0% to 1%. It was
set to 0% because the disk_usage test appeared to still be having
issues in the merge script. However, current builds appear fine:
https://logs.chromium.org/logs/chromium/buildbucket/cr-buildbucket/8801844364361381953/+/u/disk_usage_tast_test__with_patch__on_Ubuntu-18.04/raw_io.output_Merge_script_log_

But this also keeps it at 1% for win10_chromium_rel_ng. This is the
only bot seen to have failures with the experiment. This query (http://shortn/_vSOIqaHiQ9) has some false positives, but it does show
some real builds are having issue, eg:
https://ci.chromium.org/ui/p/chromium/builders/try/win10_chromium_x64_rel_ng/1314497/overview

Which appears to be a merge script failure when chromedriver_py_tests
fails on win10. Ideally the build continues on in such a case. So this
keeps the experiment low on that builder while its underlying issue
is resolved.

Bug: 1346781
Change-Id: Ia7cfbd7fad3a2bad7e84ed7951839a36b1ed1baa
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3919953
Reviewed-by: Garrett Beaty <gbeaty@google.com>
Commit-Queue: Ben Pastene <bpastene@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052088}