[a11y] refactor code for retaining AXContext and cache integration
Before this CL, RetainAXContextForDocument would also
register the current agent with the current instance of the
AXObjectCache in addition to retaining the context. This
CL refactors the RetainAXContextForDocument to be only concerned
with retaining AXContexts and creates a new helper method
AttachToAXObjectCache that is expected to perform all steps
necessary to configure the current AXObjectCache and AXContext
properly. This refactoring also fixes a bug when the a11y agent
will be added to the cache instance but there will be multiple cache
instances for the lifetime of a document (in that case, newer cache instances won't have the agent added).

LOW_COVERAGE_REASON=coverage is provided by inspector-protocol tests as instantiating agents for unit tests is somewhat involved and there are not current unit tests

Change-Id: I17ada542efa28ee640cd064753dc9ee1c9097d1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3913505
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Andrey Kosyakov <caseq@chromium.org>
Auto-Submit: Alex Rudenko <alexrudenko@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052093}