Report lock and unlock events
This changes introduces the collection and reporting of successful lock attempts and successful/failed unlock attempts.

This is only for affiliated users on managed devices and it is controlled by the policy ReportDeviceLoginLogout, the events will be surfaced on the chrome audit logs.

Design doc: go/cros-lock-unlock-dd

Bug: b:233602841
Change-Id: I553d0ec1eae755c2eb1a5cee88d3a9536cf0c98c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3832954
Reviewed-by: Ahmed Nasr <anasr@google.com>
Reviewed-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Alexander Alekseev <alemate@chromium.org>
Reviewed-by: Denis Kuznetsov <antrim@chromium.org>
Commit-Queue: Alberto Juarez <albertojuarez@google.com>
Cr-Commit-Position: refs/heads/main@{#1052137}