media/gpu/v4l2: increase input queue size
This CL increases the input queue size to be above 16, to support
test vectors with 16 reference frames.

Technically speaking the ref frames are part of the output queue,
but because of the way the code is written now, the input queue size
caps the output one. This should be fixed, but is a much larger
task.

This is verified by deploying and running the H.264 Main Profile
test vectors on asurada (MT8192) and cherry (MT8195) locally.

Bug: b/249325255, b/212179577
Change-Id: Iacd210d050084b8fd6a5d48a2be18768cf81734f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3923095
Reviewed-by: Nathan Hebert <nhebert@chromium.org>
Commit-Queue: Miguel Casas-Sanchez <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052203}