media/gpu/VDPipeline: Increase allocated frames for ImageProcessor
When ImageProcessor is needed, only 5 frames are allocated in a video
frame pool. Since rendering+displaying often blocks decoding process,
this should be larger to make less frames blocked in image processor.
The new value is the same as the number of allocated buffers in the
case when an image processor is not used.

Bug: b:249184286
Test: Google Meet call on asurada and trogdor
Test: video_decode_accelerator_tests on asurada and trogdor
Change-Id: I1daf09a665a083e10bc656e0d0e721640170c374
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3920439
Reviewed-by: Andres Calderon Jaramillo <andrescj@chromium.org>
Commit-Queue: Hirokazu Honda <hiroh@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052213}