WebUI: Move webview_manager.js to only user
As part of this change, also moving the webview_manager_test to
parent_access_browsertest, and fixing it so that the asserts in the
callbacks actually run (previously, these were skipped and the test
was passing accidentally).

Also updating incorrect closure annotations, as the original target
was not hooked up correctly.

Bug: 1364309
Change-Id: I78afbde1cd19709f72c426f78a05198a6123a42a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3914719
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: Demetrios Papadopoulos <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052216}