Use kService for StorageService
This change allows the StorageService to live inside the kService
sandbox on non-linuxy platforms. This is slightly tighter than
kUtility, for instance enabling win32 lockdown on Windows.

To allow this we have to avoid an api inside FilePath that calls
localised win32 apis and instead do ascii-style comparison - this
is ok as the compared-to file extension is fixed in the code.

Bug: 696635,1366598

Change-Id: I684b873bdbfd728287172b761b62e350cdc6dd7e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3906442
Reviewed-by: Marijn Kruisselbrink <mek@chromium.org>
Reviewed-by: Ayu Ishii <ayui@chromium.org>
Commit-Queue: Alex Gough <ajgo@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1052290}