[linux/wayland] Fixed the inconsistency for the wl_output v4 handlers.
The handlers for the methods introduced in wl_output v4 depend on the
version of Wayland, but the maximum version that the interface is bound
was not consistent with that, which could cause crashes in certain case
(see the linked crbug).  This patch fixes the issue.

Bug: 1368783
Change-Id: I2a9af5e17d1e9074b66e028d22be403aaaa2e80b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3925455
Auto-Submit: Alexander Dunaev <adunaev@igalia.com>
Commit-Queue: Nick Yamane <nickdiego@igalia.com>
Reviewed-by: Nick Yamane <nickdiego@igalia.com>
Cr-Commit-Position: refs/heads/main@{#1052310}